ogr2csv - merge CSVReader and TableReader (CSVReader removed) and fix string escaping for GDAL 2.4.0 (refs #110)
