+++ validator-core/src/main/java/fr/ign/validator/tools/TableReader.java
@@ -7,6 +7,9 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
+import org.apache.commons.csv.CSVFormat;
+import org.apache.commons.csv.CSVParser;
+import org.apache.commons.csv.CSVRecord;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -34,7 +37,7 @@ public class TableReader implements Iterator< String[] >{
 	/**
 	 * CSV file reader
 	 */
-	private CSVReader csvReader ;
+	private Iterator<CSVRecord> iterator ;
 	/**
 	 * File header
 	 */
@@ -50,10 +53,8 @@ public class TableReader implements Iterator< String[] >{
 	 * @throws IOException 
 	 */
 	private TableReader(File csvFile, Charset charset) throws IOException{
-		/*
-		 * opening file
-		 */
-		csvReader = new CSVReader(csvFile, charset);
+		CSVParser parser = CSVParser.parse(csvFile, charset, CSVFormat.RFC4180) ;
+		this.iterator = parser.iterator() ;
 		readHeader();
 	}
 
@@ -67,10 +68,10 @@ public class TableReader implements Iterator< String[] >{
 	 * @throws IOException 
 	 */
 	private void readHeader() throws IOException{
-		if ( ! csvReader.hasNext() ){
+		if ( ! hasNext() ){
 			throw new IOException("Impossible de lire l'entête");
 		}
-		String[] fields = csvReader.next() ;
+		String[] fields = next() ;
 		List<String> filteredFields = new ArrayList<String>();
 		for (String field : fields) {
 			if ( field == null || field.isEmpty() ){
@@ -80,8 +81,7 @@ public class TableReader implements Iterator< String[] >{
 		}
 		header = filteredFields.toArray(new String[filteredFields.size()]);
 	}
-	
-	
+
 	/**
 	 * @return the header
 	 */
@@ -89,28 +89,61 @@ public class TableReader implements Iterator< String[] >{
 		return header;
 	}
 
-	/**
-	 * @param header the header to set
-	 */
-	public void setHeader(String[] header) {
-		this.header = header;
-	}
-	
+
 	@Override
 	public boolean hasNext() {
-		return csvReader.hasNext() ;
+		return iterator.hasNext() ;
 	}
 
 	@Override
 	public String[] next() {
-		return csvReader.next() ;
+		CSVRecord row = iterator.next() ;
+		return toArray(row);
 	}
 
 	@Override
 	public void remove() {
-		csvReader.remove();
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * 
+	 * @param row
+	 * @return
+	 */
+	private String[] toArray(CSVRecord row) {
+		String[] result = new String[row.size()];
+		for ( int i = 0; i < row.size(); i++ ){
+			result[i] = nullifyEmptyString( trimString( row.get(i) ) ) ;
+		}
+		return result ;
+	}
+
+	/**
+	 * Trims a string
+	 * @param value
+	 * @return
+	 */
+	private String trimString(String value){
+		if ( null == value ){
+			return null ;
+		}
+		return value.trim();
 	}
 	
+	
+	/**
+	 * Converts to NULL an empty string
+	 * @param value
+	 */
+	private String nullifyEmptyString(String value){
+		if ( null == value || value.isEmpty() ){
+			return null ;
+		}else{
+			return value ;
+		}
+	}
+
 	/**
 	 * Finds the position of a column by its name in header
 	 *
